+++ crates/stdto_core/src/lib.rs
@@ -7,6 +7,8 @@ pub mod error;
 pub extern crate digest;
 #[cfg(feature = "serde")]
 pub extern crate serde;
+#[cfg(feature = "json")]
+pub extern crate serde_json;
 
 #[cfg(feature = "bytes")]
 pub use crate::{
@@ -16,6 +18,9 @@ pub use crate::{
 
 #[cfg(feature = "hash")]
 pub use crate::traits::ToHash;
+
+#[cfg(feature = "json")]
+pub use crate::traits::ToJson;
 
 #[cfg(feature = "hex")]
 pub use crate::traits::ToHex;